cleaning up fixes ukesm1_0_ll and hadgem3_gc31_ll
