+++ esmvalcore/cmor/_fixes/cmip6/hadgem3_gc31_ll.py
@@ -1,33 +1,9 @@
 """Fixes for CMIP6 HadGEM-GC31-LL."""
-from ..fix import Fix
-from .ukesm1_0_ll import Cl as BaseCl
+from .ukesm1_0_ll import Cl as BaseCl, AllVars as BaseAllVars
 
 
-class AllVars(Fix):
+class AllVars(BaseAllVars):
     """Fixes for all vars."""
-
-    def fix_metadata(self, cubes):
-        """
-        Fix parent time units.
-
-        Parameters
-        ----------
-        cube : iris.cube.CubeList
-
-        Returns
-        -------
-        iris.cube.Cube
-
-        """
-        parent_units = 'parent_time_units'
-        bad_value = 'days since 1850-01-01-00-00-00'
-        for cube in cubes:
-            try:
-                if cube.attributes[parent_units] == bad_value:
-                    cube.attributes[parent_units] = 'days since 1850-01-01'
-            except AttributeError:
-                pass
-        return cubes
 
 
 class Cl(BaseCl):
