+++ esmvalcore/cmor/_fixes/cmip6/ukesm1_0_ll.py
@@ -3,11 +3,33 @@ import iris
 
 from ..fix import Fix
 from ..shared import add_pressure_level_coordinate, fix_bounds
-from .hadgem3_gc31_ll import AllVars as BaseAllVars
 
 
-class AllVars(BaseAllVars):
+class AllVars(Fix):
     """Fixes for all vars."""
+
+    def fix_metadata(self, cubes):
+        """
+        Fix parent time units.
+
+        Parameters
+        ----------
+        cube : iris.cube.CubeList
+
+        Returns
+        -------
+        iris.cube.Cube
+
+        """
+        parent_units = 'parent_time_units'
+        bad_value = 'days since 1850-01-01-00-00-00'
+        for cube in cubes:
+            try:
+                if cube.attributes[parent_units] == bad_value:
+                    cube.attributes[parent_units] = 'days since 1850-01-01'
+            except AttributeError:
+                pass
+        return cubes
 
 
 class Cl(Fix):
